fake-sensor/src/requirements.txt
pandas
argparse
paho.mqtt.client